refactor: School 도메인 리팩토링
유저와 연결 끊도록 설정
