+++ infrastructure/mysql/src/main/kotlin/com/wespot/school/SchoolJpaEntity.kt
@@ -1,6 +1,5 @@
 package com.wespot.school
 
-import com.wespot.user.entity.UserJpaEntity
 import jakarta.persistence.*
 import org.jetbrains.annotations.NotNull
 
@@ -11,10 +10,6 @@ class SchoolJpaEntity(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long,
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "users_id", foreignKey = ForeignKey(name = "fk_school_users_id"))
-    val user: UserJpaEntity,
 
     @field:NotNull
     val name: String,
