+++ infrastructure/mysql/src/main/kotlin/com/wespot/user/entity/ProfileJpaEntity.kt
@@ -4,7 +4,7 @@ import jakarta.persistence.*
 import org.jetbrains.annotations.NotNull
 
 @Entity
-@Table(name = "fcm")
+@Table(name = "profile")
 class ProfileJpaEntity(
 
     @Id
